Files moved to respective MVC Architecture.
