+++ src/main/java/dao/Logout.java
@@ -1,4 +1,4 @@
-package DAO;
+package dao;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
