+++ src/main/java/service/DBConnection.java
@@ -1,4 +1,4 @@
-package DAO;
+package service;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
