+++ src/main/java/service/MailAuth.java
@@ -1,4 +1,4 @@
-package DAO;
+package service;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
